pyproject.toml
[tool.poetry]
name = "apdigital"
version = "0.1.0"
description = ""
authors = ["Shitikov Kirill <sh.kiruh@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.4"
django = "^2.2"
django-webpack-loader = "^0.6.0"
djangorestframework = "^3.10"
Pillow = "^6.1"

[tool.poetry.dev-dependencies]
pytest = "^3.5"
